agent: Replace prctl crate by of capctl one
While trying to get the kata-agent statically linked to the GNU libc,
we've ended up facing some issues as https://docs.rs/prctl/1.0.0/prctl/
pulls in a "-Wl,-Bdynamic" "-lc" argument that's passed to the linker,
causing then the build failure.

This seems to be caused by the following piece of code:
```
 #[link(name="c")]
 extern {
     fn prctl(option: c_int, arg2: c_ulong, arg3: c_ulong, arg4: c_ulong, arg5: c_ulong) -> c_int;
 }
```
https://github.com/viraptor/prctl-rs/blob/b86a3aad4e655e0d90c8b114fc12b4704f463937/src/lib.rs#L228-L231

Moreover, it's been found out that that crate hasn't been maintained
since 2015 and that capctl one (https://docs.rs/capctl/0.2.0/capctl/) is
a good 1:1 replacement for what we need.

Fixes: #1055

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>